requirements.txt
Flask>=3.0.0
gunicorn
spacy>=3.8.0
requests
python-dotenv
